Create web gesign
Web design is the process of creating the visual layout, structure, and usability of a website. It encompasses several disciplines and skills, including graphic design, user interface (UI) design, user experience (UX) design, and front-end web development. The main goal of web design is to create a website that is not only visually appealing but also functional, easy to navigate, and optimized for performance.